Always pack glyphs as 16*16px during codegen
The old blitstr v1 glyphs used a header word to encode the width,
height, and y-offset of trimmed and packed pixel data. This new
way just uses 16*16px sprites with the glyphs aligned in the top
left corner (left whitespace trimmed, but not top whitespace).

Using a consistent size for the sprites makes it easier to index
into the glyph data array and easier to prepare the sprites for
blitting.